issue=#667 functional-testing-support-single-case

+++ ft_test.sh
@@ -1,3 +1,17 @@
+#!/bin/bash
+
+function usage() {
+    echo "usage: $0 [case]"
+    echo "example: $0                          # all test cases"
+    echo "         $0 testcase/test_put_get.py # specify a case"
+}
+
+if [[ $# -gt 1 ]]; then
+    echo "0 or 1 command line argument"
+    usage
+    exit 1
+fi
+
 set -x -e
 
 rm -rf tmp
@@ -21,6 +35,11 @@ sh launch_tera.sh
 sleep 2
 
 export PYTHONPATH=$PYTHONPATH:../../thirdparty/include/; export PATH=$PATH:../../thirdparty/bin/
-nosetests -s -v -x > ../log/test.log
+
+if [[ $# == 0 ]]; then
+    nosetests -s -v -x > ../log/test.log
+else
+    nosetests -s -v -x $1 > ../log/test.log
+fi
 
 sh kill_tera.sh
